+++ apps/meteor/client/sidebar/header/hooks/useUserMenu.tsx
@@ -1,18 +1,20 @@
 import type { IUser } from '@rocket.chat/core-typings';
 import { useMutableCallback } from '@rocket.chat/fuselage-hooks';
+import type { GenericMenuItemProps } from '@rocket.chat/ui-client';
 import { useLogout, useTranslation } from '@rocket.chat/ui-contexts';
 import React from 'react';
 
-import type { GenericMenuItemProps } from '../../../components/GenericMenu/GenericMenuItem';
 import UserMenuHeader from '../UserMenuHeader';
 import { useAccountItems } from './useAccountItems';
 import { useStatusItems } from './useStatusItems';
+import useVoipItems from './useVoipItems';
 
 export const useUserMenu = (user: IUser) => {
 	const t = useTranslation();
 
 	const statusItems = useStatusItems();
 	const accountItems = useAccountItems();
+	const voipItems = useVoipItems();
 
 	const logout = useLogout();
 	const handleLogout = useMutableCallback(() => {
@@ -34,6 +36,9 @@ export const useUserMenu = (user: IUser) => {
 		{
 			title: t('Status'),
 			items: statusItems,
+		},
+		{
+			items: voipItems,
 		},
 		{
 			title: t('Account'),