+++ apps/meteor/client/sidebarv2/Item/Condensed.tsx
@@ -1,11 +1,11 @@
-import { IconButton, Sidebar } from '@rocket.chat/fuselage';
+import { IconButton, SidebarV2Item, SidebarV2ItemAvatarWrapper, SidebarV2ItemMenu, SidebarV2ItemTitle } from '@rocket.chat/fuselage';
 import { useEffectEvent, usePrefersReducedMotion } from '@rocket.chat/fuselage-hooks';
 import type { Keys as IconName } from '@rocket.chat/icons';
 import type { ReactElement } from 'react';
 import React, { memo, useState } from 'react';
 
 type CondensedProps = {
-	title: ReactElement | string;
+	title: string;
 	titleIcon?: ReactElement;
 	avatar: ReactElement | boolean;
 	icon?: IconName;
@@ -19,7 +19,7 @@ type CondensedProps = {
 	clickable?: boolean;
 };
 
-const Condensed = ({ icon, title = '', avatar, actions, href, unread, menu, badges, ...props }: CondensedProps) => {
+const Condensed = ({ icon, title, avatar, actions, href, unread, menu, badges, selected }: CondensedProps) => {
 	const [menuVisibility, setMenuVisibility] = useState(!!window.DISABLE_ANIMATION);
 
 	const isReduceMotionEnabled = usePrefersReducedMotion();
@@ -32,28 +32,18 @@ const Condensed = ({ icon, title = '', avatar, actions, href, unread, menu, badg
 	};
 
 	return (
-		<Sidebar.Item {...props} {...({ href } as any)} clickable={!!href}>
-			{avatar && <Sidebar.Item.Avatar>{avatar}</Sidebar.Item.Avatar>}
-			<Sidebar.Item.Content>
-				<Sidebar.Item.Wrapper>
-					{icon}
-					<Sidebar.Item.Title data-qa='sidebar-item-title' className={(unread && 'rcx-sidebar-item--highlighted') as string}>
-						{title}
-					</Sidebar.Item.Title>
-				</Sidebar.Item.Wrapper>
-				{badges && <Sidebar.Item.Badge>{badges}</Sidebar.Item.Badge>}
-				{menu && (
-					<Sidebar.Item.Menu {...handleMenuEvent}>
-						{menuVisibility ? menu() : <IconButton tabIndex={-1} aria-hidden mini rcx-sidebar-item__menu icon='kebab' />}
-					</Sidebar.Item.Menu>
-				)}
-			</Sidebar.Item.Content>
-			{actions && (
-				<Sidebar.Item.Container>
-					<Sidebar.Item.Actions>{actions}</Sidebar.Item.Actions>
-				</Sidebar.Item.Container>
+		<SidebarV2Item href={href} selected={selected}>
+			{avatar && <SidebarV2ItemAvatarWrapper>{avatar}</SidebarV2ItemAvatarWrapper>}
+			{icon && icon}
+			<SidebarV2ItemTitle unread={unread}>{title}</SidebarV2ItemTitle>
+			{badges && badges}
+			{actions && actions}
+			{menu && (
+				<SidebarV2ItemMenu {...handleMenuEvent}>
+					{menuVisibility ? menu() : <IconButton tabIndex={-1} aria-hidden mini rcx-sidebar-v2-item__menu icon='kebab' />}
+				</SidebarV2ItemMenu>
 			)}
-		</Sidebar.Item>
+		</SidebarV2Item>
 	);
 };