+++ apps/meteor/client/views/omnichannel/realTimeMonitoring/charts/ChatsPerAgentChart.js
@@ -1,5 +1,5 @@
-import { useTranslation } from '@rocket.chat/ui-contexts';
 import React, { useRef, useEffect } from 'react';
+import { useTranslation } from 'react-i18next';
 
 import { drawLineChart } from '../../../../../app/livechat/client/lib/chartHandler';
 import { AsyncStatePhase } from '../../../../hooks/useAsyncState';
@@ -20,7 +20,7 @@ const init = (canvas, context, t) =>
 	});
 
 const ChatsPerAgentChart = ({ params, reloadRef, ...props }) => {
-	const t = useTranslation();
+	const { t } = useTranslation();
 
 	const canvas = useRef();
 	const context = useRef();
@@ -56,7 +56,7 @@ const ChatsPerAgentChart = ({ params, reloadRef, ...props }) => {
 		}
 	}, [chartData, state, t, updateChartData]);
 
-	return <Chart ref={canvas} {...props} />;
+	return <Chart canvasRef={canvas} {...props} />;
 };
 
 export default ChatsPerAgentChart;