+++ apps/meteor/tests/data/livechat/department.ts
@@ -16,12 +16,16 @@ const NewDepartmentData = ((): Partial<ILivechatDepartment> => ({
 	showOnOfflineForm: true,
 }))();
 
-export const createDepartment = async (departmentData: Partial<ILivechatDepartment> = NewDepartmentData): Promise<ILivechatDepartment> => {
+export const createDepartment = async (
+	departmentData: Partial<ILivechatDepartment> = NewDepartmentData,
+	agents?: { agentId: string; count?: string; order?: string }[],
+): Promise<ILivechatDepartment> => {
 	const response = await request
 		.post(api('livechat/department'))
 		.set(credentials)
 		.send({
 			department: departmentData,
+			...(agents && { agents }),
 		})
 		.expect(200);
 	return response.body.department;
@@ -38,36 +42,44 @@ const updateDepartment = async (departmentId: string, departmentData: Partial<Li
 	return response.body.department;
 };
 
-const createDepartmentWithMethod = (
-	initialAgents: { agentId: string; username: string }[] = [],
-	{
-		allowReceiveForwardOffline = false,
-		fallbackForwardDepartment,
-	}: {
-		allowReceiveForwardOffline?: boolean;
-		fallbackForwardDepartment?: string;
-	} = {},
-) =>
+export const createDepartmentWithMethod = ({
+	initialAgents = [],
+	allowReceiveForwardOffline = false,
+	fallbackForwardDepartment,
+	name,
+	departmentUnit,
+	userCredentials = credentials,
+	departmentId = '',
+}: {
+	initialAgents?: { agentId: string; username: string }[];
+	allowReceiveForwardOffline?: boolean;
+	fallbackForwardDepartment?: string;
+	name?: string;
+	departmentUnit?: { _id?: string };
+	userCredentials?: Credentials;
+	departmentId?: string;
+} = {}): Promise<ILivechatDepartment> =>
 	new Promise((resolve, reject) => {
 		void request
 			.post(methodCall('livechat:saveDepartment'))
-			.set(credentials)
+			.set(userCredentials)
 			.send({
 				message: JSON.stringify({
 					method: 'livechat:saveDepartment',
 					params: [
-						'',
+						departmentId,
 						{
 							enabled: true,
 							email: faker.internet.email(),
 							showOnRegistration: true,
 							showOnOfflineForm: true,
-							name: `new department ${Date.now()}`,
+							name: name || `new department ${Date.now()}`,
 							description: 'created from api',
 							allowReceiveForwardOffline,
 							fallbackForwardDepartment,
 						},
 						initialAgents,
+						departmentUnit,
 					],
 					id: 'id',
 					msg: 'method',
@@ -89,7 +101,7 @@ type OnlineAgent = {
 export const createDepartmentWithAnOnlineAgent = async (): Promise<{ department: ILivechatDepartment; agent: OnlineAgent }> => {
 	const { user, credentials } = await createAnOnlineAgent();
 
-	const department = (await createDepartmentWithMethod()) as ILivechatDepartment;
+	const department = await createDepartmentWithMethod();
 
 	await addOrRemoveAgentFromDepartment(department._id, { agentId: user._id, username: user.username }, true);
 
@@ -104,7 +116,7 @@ export const createDepartmentWithAnOnlineAgent = async (): Promise<{ department:
 
 export const createDepartmentWithAgent = async (agent: OnlineAgent): Promise<{ department: ILivechatDepartment; agent: OnlineAgent }> => {
 	const { user, credentials } = agent;
-	const department = (await createDepartmentWithMethod()) as ILivechatDepartment;
+	const department = await createDepartmentWithMethod();
 
 	await addOrRemoveAgentFromDepartment(department._id, { agentId: user._id, username: user.username }, true);
 
@@ -149,7 +161,7 @@ export const createDepartmentWithAnOfflineAgent = async ({
 }> => {
 	const { user, credentials } = await createAnOfflineAgent();
 
-	const department = (await createDepartmentWithMethod(undefined, {
+	const department = (await createDepartmentWithMethod({
 		allowReceiveForwardOffline,
 		fallbackForwardDepartment,
 	})) as ILivechatDepartment;