+++ packages/livechat/src/store/index.tsx
@@ -58,6 +58,8 @@ export type StoreState = {
 			hiddenSystemMessages?: LivechatHiddenSytemMessageType[];
 			hideWatermark?: boolean;
 			livechatLogo?: { url: string };
+			transcript?: boolean;
+			visitorsCanCloseChat?: boolean;
 		};
 		online?: boolean;
 		departments: Department[];